+++ src/api/into_protobuf.rs
@@ -1,0 +1,4 @@
+pub trait IntoProtobuf: Sized {
+  type Protobuf;
+  fn into_protobuf(self) -> Self::Protobuf;
+}
