+++ src/common.rs
@@ -5,14 +5,21 @@ pub use std::{
   fmt::Debug,
   fs, io, net,
   path::{Path, PathBuf},
-  sync::Mutex,
+  sync::{Mutex, Once, ONCE_INIT},
   thread,
 };
 
+pub use rand::random;
+
+pub use api::{self, NodeId};
 pub use error::Error;
 pub use library::Library;
 pub use node::Node;
 pub use platform::{Platform, PlatformInterface};
+pub use pubkey::{self, Pubkey};
 
 #[cfg(test)]
-pub use testing::running_on_appveyor;
+pub use testing::{running_on_appveyor, test_init};
+
+#[cfg(test)]
+pub use assert_fs::{prelude::*, TempDir};
