+++ src/library.rs
@@ -1,6 +1,10 @@
 use common::*;
 
-use rusqlite::{self, types::FromSql, Connection, Row};
+use rusqlite::{
+  self,
+  types::{FromSql, ToSql},
+  Connection, Row,
+};
 
 const SQLITE_DATABASE_APPLICATION_ID: i32 = 0x1337_0000;
 
@@ -65,16 +69,16 @@ impl Library {
       }
     }
 
-    let pre_existing = database_path.exists();
+    let new_database = !database_path.exists();
 
-    if pre_existing {
+    if new_database {
       info!(
-        "Opening pre-existing library database at `{}`...",
+        "Creating new library database at `{}`...",
         database_path.display()
       );
     } else {
       info!(
-        "Creating new library database at `{}`...",
+        "Opening pre-existing library database at `{}`...",
         database_path.display()
       );
     }
@@ -91,14 +95,29 @@ impl Library {
       database_path,
     };
 
-    if pre_existing {
-      library.check_application_id()?;
+    if new_database {
+      library.initialize_database()?;
     } else {
-      library.set_application_id()?;
-      library.set_journal_mode()?;
+      library.check_database()?;
     }
 
     Ok(library)
+  }
+
+  fn initialize_database(&self) -> Result<(), Error> {
+    self.set_application_id()?;
+    self.set_journal_mode()?;
+    self.initialize_settings()?;
+
+    // library.create_collections_table()?;
+    // library.create_tables()?;
+    Ok(())
+  }
+
+  fn check_database(&self) -> Result<(), Error> {
+    self.check_application_id()?;
+    self.node_id()?;
+    Ok(())
   }
 
   fn application_id(&self) -> Result<i32, Error> {
@@ -121,14 +140,37 @@ impl Library {
     Ok(())
   }
 
-  fn execute(&self, statement: &str) -> Result<(), Error> {
+  fn initialize_settings(&self) -> Result<(), Error> {
+    self.execute("CREATE TABLE settings (node_id BLOB NOT NULL);")?;
+
+    let node_id = api::NodeId::from_pubkey(random());
+
+    let blob: &[u8] = &node_id.key().bytes;
+
+    self.call("INSERT INTO settings (node_id) VALUES (?1)", &[&blob])?;
+
+    Ok(())
+  }
+
+  fn node_id(&self) -> Result<NodeId, Error> {
+    let blob = self.query_scalar::<Vec<u8>>("SELECT node_id FROM settings;")?;
+    Pubkey::from_slice(&blob)
+      .map_err(|pubkey_error| Error::LibraryStoredNodeId { pubkey_error })
+      .map(NodeId::from_pubkey)
+  }
+
+  fn call(&self, statement: &str, params: &[&ToSql]) -> Result<(), Error> {
     self
       .connection
       .lock()
       .expect("library connection lock poisoned")
-      .execute(statement, &[])
+      .execute(statement, params)
       .embellish(self, statement)
       .map(|_| ())
+  }
+
+  fn execute(&self, statement: &str) -> Result<(), Error> {
+    self.call(statement, &[])
   }
 
   fn query_scalar<T: FromSql>(&self, statement: &str) -> Result<T, Error> {
@@ -177,12 +219,28 @@ impl<T> SqliteResultExt<T> for Result<T, rusqlite::Error> {
 }
 
 #[cfg(test)]
-mod test {
+mod tests {
   use super::*;
 
-  use assert_fs::prelude::*;
-  use assert_fs::TempDir;
   use predicates::prelude::*;
+
+  struct Test {
+    _tempdir: TempDir,
+    library: Library,
+  }
+
+  impl Test {
+    fn new() -> Test {
+      let tempdir = TempDir::new().unwrap();
+      let db = tempdir.child("library.db");
+      let library = Library::with_path(db.path()).unwrap();
+
+      Test {
+        _tempdir: tempdir,
+        library,
+      }
+    }
+  }
 
   #[test]
   fn base_path_correct_suffix() {
@@ -217,14 +275,12 @@ mod test {
   }
 
   #[test]
-  fn creation() {
-    let tempdir = TempDir::new().unwrap();
+  fn initialization() {
+    let Test { library, _tempdir } = Test::new();
 
-    let db = tempdir.child("hello.db");
+    assert!(library.database_path.is_file());
 
-    let library = Library::with_path(db.path()).unwrap();
-
-    db.assert(predicate::path::is_file());
+    library.node_id().unwrap();
 
     assert_eq!(
       library
@@ -232,6 +288,22 @@ mod test {
         .unwrap(),
       "wal"
     );
+  }
+
+  #[test]
+  fn bad_pubkey() {
+    let Test { library, _tempdir } = Test::new();
+
+    library
+      .execute("UPDATE settings SET node_id = x'012345';")
+      .unwrap();
+
+    match library.node_id() {
+      Err(Error::LibraryStoredNodeId { pubkey_error }) => {
+        assert_eq!(pubkey_error, pubkey::Error::Length { length: 3 })
+      }
+      otherwise => panic!("unexpected result: {:?}", otherwise),
+    }
   }
 
   #[test]
